+++ src/core/simulated-player/manager.ts
@@ -1,6 +1,6 @@
 import { PIDManager, type PID } from "../pid";
 import { Test, type SimulatedPlayer } from "@minecraft/server-gametest";
-import SIGN from "@/constants/YumeSignEnum";
+import { SIGN } from "@/constants";
 import { SimulatedPlayerNotFoundError, NotReadyError } from "./errors";
 import type { Entity } from "@minecraft/server";
 import type { AddSimulatedPlayerOptions, SpawnSimulatedPlayerOptions } from "./types";
