+++ src/features/commands/handlers/inventory/inventory-swap.ts
@@ -1,5 +1,5 @@
 import { commandManager } from "@/core/command";
-import { getSimPlayer } from "@/core/queries/Util";
+import { getSimPlayer } from "@/core/queries";
 import type { SimulatedPlayer } from "@minecraft/server-gametest";
 
 commandManager.registerCommand(['假人背包交换','假人交换背包'], ({entity,isEntity,sim}) => {
