+++ src/features/commands/handlers/tps/indicator.ts
@@ -1,7 +1,7 @@
 import { world } from "@minecraft/server";
 import { TPSMonitor } from "@/core/tps";
 import { commandManager } from "@/core/command";
-import { playerReady } from "@/features/events/player-ready";
+import { playerReady } from "@/features/events";
 
 const TPS_TAG = 'tps';
